Migrate utils to nosetests
